:recycle: [pkg/sync] Split Closer tests.
Now split into black-box and white-box tests.
